requirements.txt
pip install numpy pandas psycopg2-binary
pip freeze > requirements.txt

# ─── Core ───────────────────────────────────────────────────
numpy>=1.26,<3.0
pandas>=2.1,<3.0

# ─── Database ───────────────────────────────────────────────
SQLAlchemy>=2.0,<3.0
psycopg2-binary>=2.9

# ─── Synthetic data generation ──────────────────────────────
smartnoise-synth>=1.0.6
sdv~=1.35.1

# ─── ML / Privacy evaluation ────────────────────────────────
scikit-learn>=1.4,<2.0
scipy>=1.12

torch~=2.11.0
opacus~=0.14.0
sdmetrics~=0.28.0
PyYAML~=6.0.3
pydantic~=2.12.5
fastapi~=0.135.3
pydantic-settings~=2.13.1
pytest~=9.0.3